+++ src/functions/public/Add-PsCommand.ps1
@@ -46,14 +46,14 @@ function Add-PsCommand {
         [string]$WhoAmI = (whoami.EXE),
 
         # Hashtable of log messages for Write-LogMsg (can be thread-safe if a synchronized hashtable is provided)
-        [hashtable]$LogMsgCache = $Global:LogMessages
+        [hashtable]$LogBuffer = $Global:LogMessages
 
     )
 
     begin {
 
         $LogParams = @{
-            LogMsgCache  = $LogMsgCache
+            LogBuffer  = $LogBuffer
             ThisHostname = $TodaysHostname
             Type         = $DebugOutputStream
             WhoAmI       = $WhoAmI
@@ -63,7 +63,7 @@ function Add-PsCommand {
             DebugOutputStream = $DebugOutputStream
             TodaysHostname    = $TodaysHostname
             WhoAmI            = $WhoAmI
-            LogMsgCache       = $LogMsgCache
+            LogBuffer       = $LogBuffer
         }
 
         if ($CommandInfo -eq $null) {