+++ src/functions/public/Add-PsModule.ps1
@@ -34,14 +34,14 @@ function Add-PsModule {
         [string]$WhoAmI = (whoami.EXE),
 
         # Hashtable of log messages for Write-LogMsg (can be thread-safe if a synchronized hashtable is provided)
-        [hashtable]$LogMsgCache = $Global:LogMessages
+        [hashtable]$LogBuffer = $Global:LogMessages
 
     )
 
     begin {
 
         $LogParams = @{
-            LogMsgCache  = $LogMsgCache
+            LogBuffer  = $LogBuffer
             ThisHostname = $TodaysHostname
             Type         = $DebugOutputStream
             WhoAmI       = $WhoAmI