+++ src/functions/public/Convert-FromPsCommandInfoToString.ps1
@@ -16,7 +16,7 @@ function Convert-FromPsCommandInfoToString {
         [string]$WhoAmI = (whoami.EXE),
 
         # Hashtable of log messages for Write-LogMsg (can be thread-safe if a synchronized hashtable is provided)
-        [hashtable]$LogMsgCache = $Global:LogMessages
+        [hashtable]$LogBuffer = $Global:LogMessages
 
     )
     begin {
@@ -24,7 +24,7 @@ function Convert-FromPsCommandInfoToString {
             DebugOutputStream = $DebugOutputStream
             TodaysHostname    = $TodaysHostname
             WhoAmI            = $WhoAmI
-            LogMsgCache       = $LogMsgCache
+            LogBuffer       = $LogBuffer
         }
     }