+++ src/functions/public/Expand-PsCommandInfo.ps1
@@ -22,14 +22,14 @@ function Expand-PsCommandInfo {
         [string]$WhoAmI = (whoami.EXE),
 
         # Hashtable of log messages for Write-LogMsg (can be thread-safe if a synchronized hashtable is provided)
-        [hashtable]$LogMsgCache = $Global:LogMessages
+        [hashtable]$LogBuffer = $Global:LogMessages
     )
 
     $CommandInfoParams = @{
         DebugOutputStream = $DebugOutputStream
         TodaysHostname    = $TodaysHostname
         WhoAmI            = $WhoAmI
-        LogMsgCache       = $LogMsgCache
+        LogBuffer       = $LogBuffer
     }
 
     # Add the first object to the cache