+++ src/functions/public/Split-Thread.ps1
@@ -90,7 +90,7 @@ function Split-Thread {
         [string]$WhoAmI = (whoami.EXE),
 
         # Hashtable of log messages for Write-LogMsg (can be thread-safe if a synchronized hashtable is provided)
-        [hashtable]$LogMsgCache = ([hashtable]::Synchronized(@{})),
+        [hashtable]$LogBuffer = ([hashtable]::Synchronized(@{})),
 
         # ID of the parent progress bar under which to show progres
         [int]$ProgressParentId
@@ -100,7 +100,7 @@ function Split-Thread {
     begin {
 
         $LogParams = @{
-            LogMsgCache  = $LogMsgCache
+            LogBuffer    = $LogBuffer
             ThisHostname = $TodaysHostname
             Type         = $DebugOutputStream
             WhoAmI       = $WhoAmI
@@ -113,7 +113,7 @@ function Split-Thread {
             DebugOutputStream = $DebugOutputStream
             TodaysHostname    = $TodaysHostname
             WhoAmI            = $WhoAmI
-            LogMsgCache       = $LogMsgCache
+            LogBuffer         = $LogBuffer
         }
         $OriginalCommandInfo = Get-PsCommandInfo @CommandInfoParams -Command $Command
         Write-LogMsg @LogParams -Text " # Found 1 original PsCommandInfo for '$Command'"
@@ -200,7 +200,7 @@ function Split-Thread {
             RunspacePool         = $RunspacePool
             DebugOutputStream    = $DebugOutputStream
             WhoAmI               = $WhoAmI
-            LogMsgCache          = $LogMsgCache
+            LogBuffer            = $LogBuffer
         }
         if ($PSBoundParameters.ContainsKey('ProgressParentId')) {
             $ThreadParameters['ProgressParentId'] = $ProgressParentId
@@ -217,7 +217,7 @@ function Split-Thread {
             DebugOutputStream = $DebugOutputStream
             TodaysHostname    = $TodaysHostname
             WhoAmI            = $WhoAmI
-            LogMsgCache       = $LogMsgCache
+            LogBuffer         = $LogBuffer
         }
         if ($PSBoundParameters.ContainsKey('ProgressParentId')) {
             $ThreadParameters['ProgressParentId'] = $ProgressParentId