+++ src/functions/public/Wait-Thread.ps1
@@ -45,7 +45,7 @@ function Wait-Thread {
         [string]$WhoAmI = (whoami.EXE),
 
         # Hashtable of log messages for Write-LogMsg (can be thread-safe if a synchronized hashtable is provided)
-        [hashtable]$LogMsgCache = $Global:LogMessages,
+        [hashtable]$LogBuffer = $Global:LogMessages,
 
         # ID of the parent progress bar under which to show progres
         [int]$ProgressParentId
@@ -55,7 +55,7 @@ function Wait-Thread {
     begin {
 
         $LogParams = @{
-            LogMsgCache  = $LogMsgCache
+            LogBuffer  = $LogBuffer
             ThisHostname = $TodaysHostname
             Type         = $DebugOutputStream
             WhoAmI       = $WhoAmI