+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.117] - 2024-04-22 - rename LogMsgCache to LogBuffer
+
+## [1.0.116] - 2024-04-22 - rename LogMsgCache to LogBuffer
+
 ## [1.0.115] - 2024-04-07 - implement progressparentid
 
 ## [1.0.114] - 2024-02-02 - add comments/future notesto expand-pscommandinfo